refactor: extract extensions
