+++ lnbits/app.py
@@ -21,16 +21,16 @@ from lnbits.core.crud import (
     get_installed_extensions,
     update_installed_extension_state,
 )
-from lnbits.core.extensions.extension_manager import (
+from lnbits.core.helpers import migrate_extension_database
+from lnbits.core.services.extensions import (
     deactivate_extension,
 )
-from lnbits.core.extensions.helpers import version_parse
-from lnbits.core.helpers import migrate_extension_database
 from lnbits.core.tasks import (  # watchdog_task
     killswitch_task,
     wait_for_paid_invoices,
 )
 from lnbits.exceptions import register_exception_handlers
+from lnbits.helpers import version_parse
 from lnbits.settings import settings
 from lnbits.tasks import (
     cancel_all_tasks,
@@ -48,7 +48,7 @@ from lnbits.wallets import get_funding_source, set_funding_source
 from .commands import migrate_databases
 from .core import init_core_routers
 from .core.db import core_app_extra
-from .core.extensions.models import Extension, ExtensionMeta, InstallableExtension
+from .core.models.extensions import Extension, ExtensionMeta, InstallableExtension
 from .core.services import check_admin_settings, check_webpush_settings
 from .middleware import (
     CustomGZipMiddleware,
