+++ lnbits/core/views/generic.py
@@ -11,9 +11,9 @@ from fastapi.routing import APIRouter
 from lnurl import decode as lnurl_decode
 from pydantic.types import UUID4
 
-from lnbits.core.extensions.models import Extension, ExtensionMeta, InstallableExtension
 from lnbits.core.helpers import to_valid_user_id
 from lnbits.core.models import User
+from lnbits.core.models.extensions import Extension, ExtensionMeta, InstallableExtension
 from lnbits.core.services import create_invoice, create_user_account
 from lnbits.decorators import check_admin, check_user_exists
 from lnbits.helpers import template_renderer
